+++ autonomy/cli/helpers/chain.py
@@ -27,6 +27,7 @@ from aea.configurations.data_types import PackageId, PackageType
 from aea.configurations.loader import load_configuration_object
 from aea.crypto.base import Crypto, LedgerApi
 from aea.crypto.registries import crypto_registry, ledger_apis_registry
+from aea.helpers.base import IPFSHash
 from aea_ledger_ethereum.ethereum import EthereumApi, EthereumCrypto
 
 from autonomy.chain.base import UnitType
@@ -39,7 +40,7 @@ from autonomy.chain.exceptions import (
     ServiceDeployFailed,
     ServiceRegistrationFailed,
 )
-from autonomy.chain.metadata import publish_metadata
+from autonomy.chain.metadata import NFTHashOrPath, publish_metadata
 from autonomy.chain.mint import DEFAULT_NFT_IMAGE_HASH
 from autonomy.chain.mint import mint_component as _mint_component
 from autonomy.chain.mint import mint_service as _mint_service
@@ -70,7 +71,9 @@ def get_ledger_and_crypto_objects(
     chain_config = ChainConfigs.get(chain_type=chain_type)
     if chain_config.rpc is None:
         raise click.ClickException(
-            f"RPC cannot be `None` for chain config; chain_type={chain_type}"
+            f"RPC URL cannot be `None`, "
+            f"Please set the environment variable for {chain_type.value} chain "
+            f"using `{ChainConfigs.get_rpc_env_var(chain_type)}` environment variable"
         )
 
     if hwi and not HWI_PLUGIN_INSTALLED:
@@ -165,7 +168,8 @@ def mint_component(  # pylint: disable=too-many-arguments, too-many-locals
     package_type: PackageType,
     key: Optional[Path],
     chain_type: ChainType,
-    nft_image_hash: Optional[str] = None,
+    nft: Optional[NFTHashOrPath] = None,
+    owner: Optional[str] = None,
     password: Optional[str] = None,
     skip_hash_check: bool = False,
     use_latest_dependencies: bool = False,
@@ -199,10 +203,10 @@ def mint_component(  # pylint: disable=too-many-arguments, too-many-locals
             f"Cannot find configuration file for {package_type}"
         ) from e
 
-    if chain_type == ChainType.LOCAL and nft_image_hash is None:
-        nft_image_hash = DEFAULT_NFT_IMAGE_HASH
+    if chain_type == ChainType.LOCAL and nft is None:
+        nft = IPFSHash(DEFAULT_NFT_IMAGE_HASH)
 
-    if chain_type != ChainType.LOCAL and nft_image_hash is None:
+    if chain_type != ChainType.LOCAL and nft is None:
         raise click.ClickException(
             f"Please provide hash for NFT image to mint component on `{chain_type.value}` chain"
         )
@@ -213,10 +217,10 @@ def mint_component(  # pylint: disable=too-many-arguments, too-many-locals
         use_latest_dependencies=use_latest_dependencies,
     )
 
-    metadata_hash = publish_metadata(
-        public_id=package_configuration.public_id,
+    metadata_hash, metadata_string = publish_metadata(
+        package_id=package_configuration.package_id,
         package_path=package_path,
-        nft_image_hash=cast(str, nft_image_hash),
+        nft=cast(str, nft),
         description=package_configuration.description,
     )
 
@@ -225,6 +229,7 @@ def mint_component(  # pylint: disable=too-many-arguments, too-many-locals
             ledger_api=ledger_api,
             crypto=crypto,
             metadata_hash=metadata_hash,
+            owner=owner,
             component_type=UnitType.AGENT if is_agent else UnitType.COMPONENT,
             chain_type=chain_type,
             dependencies=dependencies,
@@ -246,6 +251,7 @@ def mint_component(  # pylint: disable=too-many-arguments, too-many-locals
     click.echo(f"\tMetadata Hash: {metadata_hash}")
     if token_id is not None:
         click.echo(f"\tToken ID: {token_id}")
+        (Path.cwd() / f"{token_id}.json").write_text(metadata_string)
     else:
         raise click.ClickException(
             "Could not verify metadata hash to retrieve the token ID"
@@ -260,7 +266,8 @@ def mint_service(  # pylint: disable=too-many-arguments, too-many-locals
     number_of_slots: int,
     cost_of_bond: int,
     threshold: int,
-    nft_image_hash: Optional[str] = None,
+    nft: Optional[NFTHashOrPath] = None,
+    owner: Optional[str] = None,
     password: Optional[str] = None,
     skip_hash_check: bool = False,
     use_latest_dependencies: bool = False,
@@ -294,10 +301,10 @@ def mint_service(  # pylint: disable=too-many-arguments, too-many-locals
             f"Cannot find configuration file for {PackageType.SERVICE}"
         ) from e
 
-    if chain_type == ChainType.LOCAL and nft_image_hash is None:
-        nft_image_hash = DEFAULT_NFT_IMAGE_HASH
+    if chain_type == ChainType.LOCAL and nft is None:
+        nft = IPFSHash(DEFAULT_NFT_IMAGE_HASH)
 
-    if chain_type != ChainType.LOCAL and nft_image_hash is None:
+    if chain_type != ChainType.LOCAL and nft is None:
         raise click.ClickException(
             f"Please provide hash for NFT image to mint component on `{chain_type.value}` chain"
         )
@@ -317,10 +324,10 @@ def mint_service(  # pylint: disable=too-many-arguments, too-many-locals
             f"\n\tAgent ID: {package_configuration.agent}"
         )
 
-    metadata_hash = publish_metadata(
-        public_id=package_configuration.public_id,
+    metadata_hash, metadata_string = publish_metadata(
+        package_id=package_configuration.package_id,
         package_path=package_path,
-        nft_image_hash=cast(str, nft_image_hash),
+        nft=cast(str, nft),
         description=package_configuration.description,
     )
 
@@ -341,6 +348,7 @@ def mint_service(  # pylint: disable=too-many-arguments, too-many-locals
             ],
             threshold=threshold,
             timeout=timeout,
+            owner=owner,
         )
     except ComponentMintFailed as e:
         raise click.ClickException(
@@ -356,13 +364,14 @@ def mint_service(  # pylint: disable=too-many-arguments, too-many-locals
     click.echo(f"\tMetadata Hash: {metadata_hash}")
     if token_id is not None:
         click.echo(f"\tToken ID: {token_id}")
+        (Path.cwd() / f"{token_id}.json").write_text(metadata_string)
     else:
         raise click.ClickException(
             "Could not verify metadata hash to retrieve the token ID"
         )
 
 
-def activate_service(
+def activate_service(  # pylint: disable=too-many-arguments
     service_id: int,
     key: Path,
     chain_type: ChainType,