+++ tests/test_autonomy/test_chain/base.py
@@ -19,6 +19,7 @@
 
 """Test tools for CLI commands that use `autonomy.chain`"""
 
+import os
 import re
 from pathlib import Path
 from typing import Dict, List, Optional
@@ -26,10 +27,12 @@ from typing import Dict, List, Optional
 import pytest
 from aea.configurations.data_types import PackageId, PackageType, PublicId
 from aea.crypto.base import Crypto, LedgerApi
+from aea.helpers.base import IPFSHash
 from aea_test_autonomy.configurations import ETHEREUM_KEY_DEPLOYER
 from aea_test_autonomy.docker.base import skip_docker_tests
 from aea_test_autonomy.fixture_helpers import registries_scope_class  # noqa: F401
 
+from autonomy.chain.base import registry_contracts
 from autonomy.chain.config import ChainType
 from autonomy.chain.constants import (
     AGENT_REGISTRY_ADDRESS_LOCAL,
@@ -43,7 +46,7 @@ from autonomy.chain.mint import (
     mint_component,
     mint_service,
 )
-from autonomy.chain.utils import resolve_component_id
+from autonomy.chain.utils import parse_public_id_from_metadata, resolve_component_id
 from autonomy.cli.helpers.chain import get_ledger_and_crypto_objects
 from autonomy.cli.packages import get_package_manager
 
@@ -133,6 +136,43 @@ class BaseChainInteractionTest(BaseCliTest):
         *_, token_id = matches
         return int(token_id)
 
+    def verify_owner_address(
+        self, token_id: int, owner: str, package_id: PackageId
+    ) -> None:
+        """Verify minted token id"""
+
+        if package_id.package_type == PackageType.SERVICE:
+            on_chain_owner = (
+                registry_contracts.service_registry.get_instance(
+                    ledger_api=self.ledger_api,
+                    contract_address=SERVICE_REGISTRY_ADDRESS_LOCAL,
+                )
+                .functions.ownerOf(token_id)
+                .call()
+            )
+        elif package_id.package_type == PackageType.AGENT:
+            on_chain_owner = (
+                registry_contracts.component_registry.get_instance(
+                    ledger_api=self.ledger_api,
+                    contract_address=AGENT_REGISTRY_ADDRESS_LOCAL,
+                )
+                .functions.ownerOf(token_id)
+                .call()
+            )
+        else:
+            on_chain_owner = (
+                registry_contracts.component_registry.get_instance(
+                    ledger_api=self.ledger_api,
+                    contract_address=COMPONENT_REGISTRY_ADDRESS_LOCAL,
+                )
+                .functions.ownerOf(token_id)
+                .call()
+            )
+
+        assert (
+            owner.lower() == on_chain_owner.lower()
+        ), f"Owner verification failed; Owner address={owner}; On chain owner address={on_chain_owner}"
+
     def verify_minted_token_id(self, token_id: int, package_id: PackageId) -> None:
         """Verify minted token id"""
 
@@ -154,11 +194,19 @@ class BaseChainInteractionTest(BaseCliTest):
             is_service=is_service,
         )
 
-        minted_public_id = PublicId.from_str(metadata["name"])
+        minted_public_id = parse_public_id_from_metadata(metadata["name"])
         assert package_id.public_id.to_any() == minted_public_id.to_any(), (
             minted_public_id,
             package_id.public_id,
         )
+
+    @staticmethod
+    def verify_and_remove_metadata_file(token_id: int) -> None:
+        """Verify and remove the metadata file."""
+
+        metadata_file = Path(f"{token_id}.json").resolve()
+        assert metadata_file.exists()
+        os.remove(metadata_file)
 
     def mint_component(
         self,
@@ -171,10 +219,10 @@ class BaseChainInteractionTest(BaseCliTest):
         package_path = DUMMY_PACKAGE_MANAGER.package_path_from_package_id(
             package_id=package_id
         )
-        metadata_hash = publish_metadata(
-            public_id=package_id.public_id,
+        metadata_hash, _ = publish_metadata(
+            package_id=package_id,
             package_path=package_path,
-            nft_image_hash=DEFAULT_NFT_IMAGE_HASH,
+            nft=IPFSHash(DEFAULT_NFT_IMAGE_HASH),
             description="Dummy package for testing",
         )
         if package_id.package_type == PackageType.SERVICE:
@@ -186,7 +234,7 @@ class BaseChainInteractionTest(BaseCliTest):
                 crypto=self.crypto,
                 metadata_hash=metadata_hash,
                 chain_type=ChainType.LOCAL,
-                **service_mint_parameters
+                **service_mint_parameters,
             )
         else:
             token_id = mint_component(