+++ tests/test_autonomy/test_cli/test_helpers/test_chain_helpers.py
@@ -50,7 +50,8 @@ DUMMY_METADATA_HASH = (
 
 
 publish_metadata_patch = mock.patch(
-    "autonomy.cli.helpers.chain.publish_metadata", return_value=DUMMY_METADATA_HASH
+    "autonomy.cli.helpers.chain.publish_metadata",
+    return_value=(DUMMY_METADATA_HASH, ""),
 )
 
 _ = registry_contracts.service_registry
@@ -118,7 +119,11 @@ class TestMintComponentMethod:
 
         with pytest.raises(
             click.ClickException,
-            match="RPC cannot be `None` for chain config; chain_type=ChainType.GOERLI",
+            match=(
+                "RPC URL cannot be `None`, "
+                "Please set the environment variable for goerli chain "
+                "using `GOERLI_CHAIN_RPC` environment variable"
+            ),
         ):
             mint_component(
                 package_path=PACKAGE_DIR,
@@ -163,7 +168,10 @@ class TestMintComponentMethod:
         ):
             with mock.patch(
                 "autonomy.cli.helpers.chain.verify_component_dependencies"
-            ), mock.patch("autonomy.cli.helpers.chain.publish_metadata"), mock.patch(
+            ), mock.patch(
+                "autonomy.cli.helpers.chain.publish_metadata",
+                return_value=(None, None),
+            ), mock.patch(
                 "autonomy.chain.mint.transact"
             ), mock.patch.object(
                 registry_contracts._registries_manager,
@@ -199,7 +207,8 @@ def test_mint_service_timeout() -> None:
         ), mock.patch(
             "autonomy.cli.helpers.chain.load_configuration_object"
         ), mock.patch(
-            "autonomy.cli.helpers.chain.publish_metadata"
+            "autonomy.cli.helpers.chain.publish_metadata",
+            return_value=(None, None),
         ), mock.patch(
             "autonomy.chain.mint.transact"
         ), mock.patch.object(